fixed the auto key in new branch, using pyautogui as a quick fix, will try to get this to backend later
